AI/requirements.txt
pandas
numpy
scikit-learn
openai
pymongo
python-dotenv
flask
flask-cors
requests